feat(mesh): close all 7 P2P gaps — anti-poisoning filter wired, karma gating, CRDT sets, near-dup, tool+LoRA validators
Mesh phase 3: wire the 9-layer anti-poisoning filter into the live packet-
receive path (was test-endpoint only), add karma trust gating, dependency-free
near-duplicate detection, CRDT G-Set/2P-Set + vector clock, tool-manifest +
LoRA-delta content validators. Plus the prior security-audit batch (11 fixes).

New: internal/mesh/{similarity,pipeline,karma_gate,toolvalidate,crdt_sets,lora}.go
     internal/mesh/mesh_phase3_test.go (9 tests)
     handlers_ssrf_guard.go (+ test), doc/AUDIT_INVENTORY.md

build + vet + test (32 ok / 0 fail) + -race CLEAN. Runtime-verified on a
running router (mesh boots, validators reject/accept correctly, 0 panic).
LoRA weight-apply honestly deferred (needs fine-tuning runtime); transport +
checksum + signature verification are live.

+++ internal/router/dispatcher.go
@@ -591,7 +591,10 @@ func pickComboModel(c *store.Combo) string {
 		return c.Models[i]
 	case store.ComboStrategyRandom:
 		// Use time.Now nanos modulo as cheap PRNG (no crypto needed for routing).
-		return c.Models[int(time.Now().UnixNano())%len(c.Models)]
+		// uint64 conversion avoids a negative index: on 32-bit int builds (or any
+		// nanos value that overflows int) `int(x) % len` can be negative, which
+		// would panic with index-out-of-range and crash the request goroutine.
+		return c.Models[int(uint64(time.Now().UnixNano())%uint64(len(c.Models)))]
 	case store.ComboStrategyCostOptimal:
 		// Pick model yang harga input+output terendah (estimateCost as proxy).
 		bestModel := c.Models[0]
